Integrate game code into Vite React project
