+++ README.md
@@ -1,15 +1,21 @@
 # Lingua Avventura
 
-Proyecto de aventura interactiva para practicar idiomas.
+SPA gamificada para aprender italiano (UI en español) con minijuego tipo Snake (ratón y queso) que narra palabras nuevas al comer.
 
-## Scripts disponibles
+## Desarrollo local
+```bash
+npm install
+npm run dev
+```
 
-- `npm run dev`: inicia el servidor de desarrollo con Vite.
-- `npm run build`: genera los archivos de producción.
-- `npm run preview`: sirve la versión generada.
-- `npm test`: ejecuta la tarea de pruebas (actualmente no hay pruebas configuradas).
+## Build
+```bash
+npm run build
+npm run preview
+```
 
-## Requisitos
-
-1. Instalar dependencias con `npm install`.
-2. Ejecutar cualquiera de los scripts anteriores según corresponda.
+## GitHub Pages
+1. Crea el repo `lingua-avventura`.
+2. Empuja el proyecto.
+3. En `vite.config.js` ajusta `base: '/lingua-avventura/'`.
+4. `npm run build` y publica el contenido de `dist/` en la rama `gh-pages` (o usa GitHub Action).
