--- game.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Inicio de la aventura</title>
-    <script src="https://cdn.tailwindcss.com"></script>
-  </head>
-  <body class="min-h-screen bg-gradient-to-br from-indigo-300 via-purple-300 to-pink-300 flex items-center justify-center">
-    <h2 class="text-3xl md:text-4xl font-semibold text-gray-800 drop-shadow-lg">
-      ¡La aventura comenzará pronto!
-    </h2>
-  </body>
-</html>
