+++ vite.config.js
@@ -1,17 +1,7 @@
-import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
-import { resolve } from 'path';
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
 
-// https://vitejs.dev/config/
 export default defineConfig({
-  base: '/lingua-avventura/', // ⚠️ usa el nombre de tu repo exacto
   plugins: [react()],
-  build: {
-    rollupOptions: {
-      input: {
-        main: resolve(__dirname, 'index.html'),
-        game: resolve(__dirname, 'game.html'),
-      },
-    },
-  },
-});
+  base: '' // <- si usas GitHub Pages con user/REPO, cambia a '/REPO/'
+})
